backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.8.3
anthropic==0.37.1
pydantic==2.9.2
